backend/requirements.txt
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
environs==10.3.0
fastapi==0.109.2
fastapi-security==0.5.0
fastapi_cors==0.0.6
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
jwt==1.3.1
kiwisolver==1.4.5
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib==3.8.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
passlib==1.7.4
pillow==10.2.0
pluggy==1.4.0
psutil==5.9.8
psycopg2==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.16.1
PyJWT==2.8.0
pyparsing==3.1.1
pytest==8.0.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.36.3
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.1.15
urllib3==2.2.1
uvicorn==0.27.0.post1
